Implemented fast consistency check, renamed variable, and moved regression tests
